move Golang to own dir
